chore: tidy Curric* components adding tests/storybooks
