+++ src/components/CurriculumComponents/CurricVisualiserFiltersModal/CurricVisualiserFiltersModal.fixtures.ts
