--- src/components/CurriculumComponents/CurriculumVisualiser/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CurriculumVisualiser";
